#435 Rename typing helper module
