+++ neclib/core/__init__.py
@@ -12,7 +12,7 @@ from . import environ  # noqa: F401
 from . import formatting  # noqa: F401
 from . import math  # noqa: F401
 from . import security  # noqa: F401
-from . import type_aliases  # noqa: F401
+from . import types  # noqa: F401
 from . import units  # noqa: F401
 from .configuration import *  # noqa: F401, F403
 from .data_type import *  # noqa: F401, F403
