+++ neclib/recorders/file_writer.py
@@ -3,7 +3,7 @@ from pathlib import Path
 from typing import Optional
 
 from ..core import get_logger
-from ..core.type_aliases import TextLike
+from ..core.types import TextLike
 from .writer_base import Writer
