+++ neclib/recorders/necstdb_writer.py
@@ -8,7 +8,7 @@ from typing import Any, Callable, Dict, Iterable, List, Optional, Sequence, Tupl
 import necstdb
 
 from ..core import get_logger
-from ..core.type_aliases import TextLike
+from ..core.types import TextLike
 from .writer_base import Writer
 
 TextData = Union[TextLike, List[TextLike]]
